Task #11 - Modified the scheduled method using an abstract class to an easy extension
