+++ src/main/java/zoeque/limitchecker/domain/repository/StoredItemRepositoryImpl.java
@@ -1,6 +1,5 @@
 package zoeque.limitchecker.domain.repository;
 
-import java.time.LocalDateTime;
 import java.util.List;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -9,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import zoeque.limitchecker.domain.entity.StoredItem;
 
 @Repository
-public interface IStoredItemRepository
+public interface StoredItemRepositoryImpl
         extends JpaRepository<StoredItem, String>,
         JpaSpecificationExecutor<StoredItem> {
   List<StoredItem> findAll(Specification<StoredItem> spec);
